Test Coverage for older django versions (#24968)
Related to this issue:
https://github.com/microsoft/vscode-python/issues/24199

@mcobalchinisoftfocus Discovered an issue with older django versions,
which didn't have the main function in the manage.py


https://github.com/microsoft/vscode-python/pull/24927#issuecomment-2779480139

I've fixed this issue by executing the code in manage.py with __name__
set to __main__ instead of relying on main function being there.

I've also adjusted the test, so that it would cover this case.